ZAVRSNI RAD ALGEBRA TKINTER/requirements.txt
numpy==1.24.2
python-dateutil==2.8.2
requests==2.28.2
SQLAlchemy==2.0.4
urllib3==1.26.14
weather-api==1.0

future==0.18.3
Pillow==9.4.0
matplotlib==3.7.0